// src/app/generated_services/model/paginationNotificationDTO.ts
/**
 * BjjClutch Backend API
 * API for managing students, belts, graduation requirements, notices, frequencies and user authentication with Keycloak
 *
 * Contact: marcosturossi@gmail.com.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShowNotificationDTO } from './showNotificationDTO';


export interface PaginationNotificationDTO { 
    items?: Array<ShowNotificationDTO> | null;
    totalCount?: number;
    pageNumber?: number;
    pageSize?: number;
    readonly totalPages?: number;
}